chatterpy-backend/requirements.txt
aiohttp==3.6.2
asgiref==3.2.7
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.9.1
boto3==1.14.40
botocore==1.17.40
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-rest-auth==1.0.9
Django==3.0.8
django-allauth==0.42.0
django-better-admin-arrayfield==1.1.0
django-bootstrap4==1.1.1
django-cors-headers==3.4.0
django-extensions==3.0.3
django-filter==2.3.0
django-storages==1.9.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
Faker==4.1.1
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
Markdown==3.2.2
MarkupSafe==1.1.1
messagebird==1.5.0
model-bakery==1.1.1
multidict==4.7.6
ndg-httpsclient==0.5.1
oauthlib==3.1.0
phonenumbers==8.12.11
Pillow==7.1.2
psycopg2-binary==2.8.5
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
signalwire==2.0.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
twilio==6.16.4
uritemplate==3.0.1
urllib3==1.25.9
yarl==1.4.2
zipp==3.1.0
